@@ -1,12 +1,11 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Boilerplate.Api.Controllers.Accounts
+namespace Boilerplate.Api.Controllers.Accounts;
+
+public class CreateAccountRequest
 {
-    public class CreateAccountRequest
-    {
-        [Required]
-        public string Email { get; set; }
-        [Required]
-        public string FullName { get; set; }
-    }
+    [Required]
+    public string Email { get; set; }
+    [Required]
+    public string FullName { get; set; }
 }
