@@ -1,12 +1,11 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Boilerplate.Api.Controllers.Accounts
+namespace Boilerplate.Api.Controllers.Accounts;
+
+public class LoginAccountRequest
 {
-    public class LoginAccountRequest
-    {
-        [Required]
-        public string Email { get; set; }
-        [Required]
-        public string Password { get; set; }
-    }
+    [Required]
+    public string Email { get; set; }
+    [Required]
+    public string Password { get; set; }
 }
