@@ -1,13 +1,12 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace Boilerplate.Api.Controllers.Accounts
+namespace Boilerplate.Api.Controllers.Accounts;
+
+public class UpdateAccountPasswordRequest
 {
-    public class UpdateAccountPasswordRequest
-    {
-        [Required]
-        public Guid ResetPasswordToken { get; set; }
-        [Required]
-        public string Password { get; set; }
-    }
+    [Required]
+    public Guid ResetPasswordToken { get; set; }
+    [Required]
+    public string Password { get; set; }
 }
