@@ -1,16 +1,15 @@
 using System;
 using Boilerplate.Api.Infrastructure.ErrorHandling;
 
-namespace Boilerplate.Api.Domain.Exceptions
+namespace Boilerplate.Api.Domain.Exceptions;
+
+/// <summary>
+/// To be thrown when business logic fails due to a conflict.
+/// This exception type is caught by the global filter and formatted before sending responses to the client.
+/// </summary>
+public class NotFoundException : BusinessRuleException
 {
-    /// <summary>
-    /// To be thrown when business logic fails due to a conflict.
-    /// This exception type is caught by the global filter and formatted before sending responses to the client.
-    /// </summary>
-    public class NotFoundException : BusinessRuleException
+    public NotFoundException(ErrorCode errorCode) : base(errorCode)
     {
-        public NotFoundException(ErrorCode errorCode) : base(errorCode)
-        {
-        }
     }
 }
