@@ -6,30 +6,29 @@ using Boilerplate.Api.Infrastructure.Database.Entities;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace Boilerplate.Api.Domain.Queries.ContentPages
+namespace Boilerplate.Api.Domain.Queries.Accounts;
+
+public static class GetAccounts
 {
-    public static class GetAccounts
+    // query
+    public record Query() : IRequest<IEnumerable<AccountEntity>>;
+
+    // handler
+    public class Handler : IRequestHandler<Query, IEnumerable<AccountEntity>>
     {
-        // query
-        public record Query() : IRequest<IEnumerable<AccountEntity>>;
+        private readonly AppDbContext _dbContext;
 
-        // handler
-        public class Handler : IRequestHandler<Query, IEnumerable<AccountEntity>>
+        public Handler(AppDbContext dbContext)
         {
-            private readonly AppDbContext _dbContext;
+            _dbContext = dbContext;
+        }
 
-            public Handler(AppDbContext dbContext)
-            {
-                _dbContext = dbContext;
-            }
-
-            public async Task<IEnumerable<AccountEntity>> Handle(Query request, CancellationToken cancellationToken)
-            {
-                var accounts = await _dbContext.Accounts
-                    .AsNoTracking()
-                    .ToArrayAsync();
-                return accounts;
-            }
+        public async Task<IEnumerable<AccountEntity>> Handle(Query request, CancellationToken cancellationToken)
+        {
+            var accounts = await _dbContext.Accounts
+                .AsNoTracking()
+                .ToArrayAsync();
+            return accounts;
         }
     }
 }
