@@ -1,10 +1,9 @@
 using System;
 using System.Threading.Tasks;
 
-namespace Boilerplate.Api.Domain.Services
+namespace Boilerplate.Api.Domain.Services;
+
+public interface IMailService
 {
-    public interface IMailService
-    {
-        Task SendResetPasswordEmail(string email, string fullName, Guid resetPasswordToken);
-    }
+    Task SendResetPasswordEmail(string email, string fullName, Guid resetPasswordToken);
 }
