@@ -2,43 +2,42 @@ using System;
 using System.Security.Cryptography;
 using Microsoft.AspNetCore.Cryptography.KeyDerivation;
 
-namespace Boilerplate.Api.Domain.Services
+namespace Boilerplate.Api.Domain.Services;
+
+public interface IPasswordService
 {
-    public interface IPasswordService
+    PasswordDto EncryptPassword(string password);
+    bool VerifyPassword(string enteredPassword, byte[] salt, string storedPassword);
+}
+
+public class PasswordService : IPasswordService
+{
+    public PasswordDto EncryptPassword(string password)
     {
-        HashSalt EncryptPassword(string password);
-        bool VerifyPassword(string enteredPassword, byte[] salt, string storedPassword);
+        byte[] salt = new byte[128 / 8]; // Generate a 128-bit salt using a secure PRNG
+        using (var rng = RandomNumberGenerator.Create())
+        {
+            rng.GetBytes(salt);
+        }
+        string encryptedpw = Convert.ToBase64String(KeyDerivation.Pbkdf2(
+            password: password,
+            salt: salt,
+            prf: KeyDerivationPrf.HMACSHA1,
+            iterationCount: 10000,
+            numBytesRequested: 256 / 8
+        ));
+        return new PasswordDto(encryptedpw, salt);
     }
 
-    public class PasswordService : IPasswordService
+    public bool VerifyPassword(string enteredPassword, byte[] salt, string storedPassword)
     {
-        public HashSalt EncryptPassword(string password)
-        {
-            byte[] salt = new byte[128 / 8]; // Generate a 128-bit salt using a secure PRNG
-            using (var rng = RandomNumberGenerator.Create())
-            {
-                rng.GetBytes(salt);
-            }
-            string encryptedpw = Convert.ToBase64String(KeyDerivation.Pbkdf2(
-                password: password,
-                salt: salt,
-                prf: KeyDerivationPrf.HMACSHA1,
-                iterationCount: 10000,
-                numBytesRequested: 256 / 8
-            ));
-            return new HashSalt(encryptedpw, salt);
-        }
-
-        public bool VerifyPassword(string enteredPassword, byte[] salt, string storedPassword)
-        {
-            string encryptedPassw = Convert.ToBase64String(KeyDerivation.Pbkdf2(
-                password: enteredPassword,
-                salt: salt,
-                prf: KeyDerivationPrf.HMACSHA1,
-                iterationCount: 10000,
-                numBytesRequested: 256 / 8
-            ));
-            return encryptedPassw == storedPassword;
-        }
+        string encryptedPassw = Convert.ToBase64String(KeyDerivation.Pbkdf2(
+            password: enteredPassword,
+            salt: salt,
+            prf: KeyDerivationPrf.HMACSHA1,
+            iterationCount: 10000,
+            numBytesRequested: 256 / 8
+        ));
+        return encryptedPassw == storedPassword;
     }
 }
