@@ -1,8 +1,7 @@
-namespace Boilerplate.Api.Infrastructure.Authorization
+namespace Boilerplate.Api.Infrastructure.Authorization;
+
+public static class CustomClaimTypes
 {
-    public static class CustomClaimTypes
-    {
-        public const string Email = "email";
-        public const string FullName = "fullName";
-    }
+    public const string Email = "email";
+    public const string FullName = "fullName";
 }
