@@ -1,14 +1,13 @@
-namespace Boilerplate.Api.Infrastructure.ErrorHandling
-{
-    public class ErrorCode
-    {
-        public ErrorCode(int code, string message)
-        {
-            Code = code;
-            Message = message;
-        }
+namespace Boilerplate.Api.Infrastructure.ErrorHandling;
 
-        public int Code { get; set; }
-        public string Message { get; set; }
+public class ErrorCode
+{
+    public ErrorCode(int code, string message)
+    {
+        Code = code;
+        Message = message;
     }
+
+    public int Code { get; set; }
+    public string Message { get; set; }
 }
