@@ -1,5 +1,4 @@
 using Boilerplate.Api.Controllers;
-using Boilerplate.Api.Domain.Exceptions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -7,40 +6,38 @@ using System.Linq;
 using System.Net.Mime;
 using System.Text.Json;
 
-namespace Boilerplate.Api.Infrastructure.ErrorHandling
+namespace Boilerplate.Api.Infrastructure.ErrorHandling;
+
+/// <summary>
+/// Catch invalid model states and format a proper response to the client
+/// </summary>
+public static class ModelStateValidationBehaviour
 {
     /// <summary>
-    /// Catch invalid model states and format a proper response to the client
+    /// handle modelstate validation errors
     /// </summary>
-    public static class ModelStateValidationBehaviour
-    {
-        /// <summary>
-        /// handle modelstate validation errors
-        /// </summary>
-        public static IMvcBuilder ConfigureApiBehaviorOptions(this IMvcBuilder builder) =>
-            builder.ConfigureApiBehaviorOptions(options =>
+    public static IMvcBuilder ConfigureApiBehaviorOptions(this IMvcBuilder builder) =>
+        builder.ConfigureApiBehaviorOptions(options =>
+        {
+            options.InvalidModelStateResponseFactory = context =>
             {
-                options.InvalidModelStateResponseFactory = context =>
-                {
-                    var errorResponse = new ApiErrorResponse(
-                        ErrorCodes.VALIDATION.Code,
-                        ErrorCodes.VALIDATION.Message,
-                        context.ModelState.Keys
-                            .SelectMany(key => context.ModelState[key].Errors
-                                .Select(x => new ModelValidationError(key, x.ErrorMessage)))
-                            .ToArray());
+                var errorResponse = new ApiErrorResponse(
+                    ErrorCodes.VALIDATION.Code,
+                    ErrorCodes.VALIDATION.Message,
+                    context.ModelState.Keys
+                        .SelectMany(key => context.ModelState[key].Errors
+                            .Select(x => new ModelValidationError(key, x.ErrorMessage)))
+                        .ToArray());
 
-                    var result = new BadRequestObjectResult(errorResponse);
-                    result.ContentTypes.Add(MediaTypeNames.Application.Json);
+                var result = new BadRequestObjectResult(errorResponse);
+                result.ContentTypes.Add(MediaTypeNames.Application.Json);
 
-                    var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<Startup>>();
-                    logger.LogWarning("ModelStateValidationError {errorResponse}", JsonSerializer.Serialize(errorResponse));
+                var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<Program>>();
+                logger.LogWarning("ModelStateValidationError {errorResponse}", JsonSerializer.Serialize(errorResponse));
 
-                    return result;
-                };
+                return result;
+            };
                 // necessary to disable the defualt ProblemDetails model in Swagger
                 options.SuppressMapClientErrors = true;
-            });
-
-    }
+        });
 }
