@@ -1,43 +1,41 @@
 using System;
-using System.IO;
-using System.Reflection;
 using System.Text;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
 
-namespace Boilerplate.Api.Infrastructure.Extensions{
+namespace Boilerplate.Api.Infrastructure.Extensions;
 
-    public static class AuthenticationExtensions
+public static class AuthenticationExtensions
+{
+    /// <summary>
+    /// Add JWT bearer scheme. 
+    /// Using the [Authorize] attribute on controllers or controller-actions, will activate bearer authentication.
+    /// </summary>
+    public static IServiceCollection AddAuth(this IServiceCollection services, AuthorizationSettings settings)
     {
-        /// <summary>
-        /// Add JWT bearer scheme. 
-        /// Using the [Authorize] attribute on controllers or controller-actions, will activate bearer authentication.
-        /// </summary>
-        public static IServiceCollection AddAuth(this IServiceCollection services, AuthorizationSettings settings){
-            services.AddAuthentication(options =>
+        services.AddAuthentication(options =>
+        {
+            options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+            options.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
+            options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+        })
+        .AddJwtBearer(cfg =>
+        {
+            cfg.RequireHttpsMetadata = false;
+            cfg.SaveToken = true;
+            cfg.TokenValidationParameters = new TokenValidationParameters
             {
-                options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                options.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-            .AddJwtBearer(cfg =>
-            {
-                cfg.RequireHttpsMetadata = false;
-                cfg.SaveToken = true;
-                cfg.TokenValidationParameters = new TokenValidationParameters
-                {
-                    ValidateIssuer = true,
-                    ValidIssuer = settings.Issuer,
-                    ValidateAudience = true,
-                    ValidAudience = settings.Audience,
-                    ValidateLifetime = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(settings.JwtKey)),
-                    ClockSkew = TimeSpan.Zero
-                };
-            });
-            services.AddAuthorization();
-            return services;
-        }
+                ValidateIssuer = true,
+                ValidIssuer = settings.Issuer,
+                ValidateAudience = true,
+                ValidAudience = settings.Audience,
+                ValidateLifetime = true,
+                IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(settings.JwtKey)),
+                ClockSkew = TimeSpan.Zero
+            };
+        });
+        services.AddAuthorization();
+        return services;
     }
 }
