@@ -5,22 +5,21 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using SendGrid;
 
-namespace Boilerplate.Api.Tests.TestUtils.Stubs
+namespace Boilerplate.Api.Tests.TestUtils.Stubs;
+
+public class STUB_Mailservice : SendGridService
 {
-    public class STUB_Mailservice : SendGridService
+    public STUB_Mailservice(
+        ILogger<SendGridService> logger,
+        IOptions<SendGridSettings> sendGridSettings,
+        ISendGridClient sendGridClient)
+        : base(logger, sendGridSettings, sendGridClient)
     {
-        public STUB_Mailservice(
-            ILogger<SendGridService> logger,
-            IOptions<SendGridSettings> sendGridSettings,
-            ISendGridClient sendGridClient)
-            : base(logger, sendGridSettings, sendGridClient)
-        {
-        }
+    }
 
 
-        protected override async Task SendEmail(SendGridMailDto mailDto)
-        {
-            await Task.Run(() => { });
-        }
+    protected override async Task SendEmail(SendGridMailDto mailDto)
+    {
+        await Task.Run(() => { });
     }
 }
